SentientStream/requirements.txt
# ── Database ──────────────────────────────
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── HTTP / Async ───────────────────────────
httpx==0.26.0
aiohttp==3.9.3
aiofiles==23.2.1
requests==2.31.0

# ── AI: Groq (Vision) ─────────────────────
groq==0.4.2

# ── AI: Embeddings (runs locally, free) ───
sentence-transformers==2.5.1
torch==2.2.0
transformers==4.37.2

# ── Video Processing ───────────────────────
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4

# ── Vector Store ───────────────────────────
faiss-cpu==1.7.4

# ── Backend ────────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ── Utilities ──────────────────────────────
python-dotenv==1.0.1
pydantic==2.6.1
tqdm==4.66.2
loguru==0.7.2
